requirements.txt
akshare>=1.11.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
jinja2>=3.1.0
plotly>=5.15.0
requests>=2.31.0
lxml>=4.9.0
fastapi>=0.116.0
uvicorn>=0.35.0
sqlalchemy>=2.0.37
apscheduler>=3.10.4
pytest>=8.3.4
